backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.55.3
feedparser==6.0.11
gunicorn==21.2.0